+++ src/CMakeLists.txt
@@ -158,6 +158,7 @@ plotting_surface.cpp
 plotting_windows.cpp
 plotting_xyouts.cpp
 plotting.cpp
+poly_2d.cpp
 print.cpp
 print_tree.cpp
 prognode.cpp
