+++ src/math_fun_jmg.cpp
@@ -892,1254 +892,5 @@ namespace lib {
       return res;
     }
   }
-//-----------------------------------------------------------------POLY_2D
-  // Function to return a^n
-
-  int powerOptimised(OMPInt a, OMPInt n) {
-
-	// Stores final answer
-	OMPInt ans = 1;
-
-	while (n > 0) {
-
-	  OMPInt last_bit = (n & 1);
-
-	  // Check if current LSB
-	  // is set
-	  if (last_bit) {
-		ans = ans * a;
-	  }
-
-	  a = a * a;
-
-	  // Right shift
-	  n = n >> 1;
-	}
-
-	return ans;
-  }
-
-  /*-------------------------------------------------------------------------*/
-  /**
-	@brief	Compute the value of a poly2d at a given point.
-	@param	p	Poly2d object.
-	@param	x	x coordinate.
-	@param	y	y coordinate.
-	@return	The value of the 2d polynomial at (x,y) as a float.
-
-	This computes the value of a poly2d in a single point. To
-	compute many values in a row, see poly2d_compute_array().
-   this version for DLong ( x and y are DLong pixels)
-   */
-
-  /*--------------------------------------------------------------------------*/
-  DFloat poly2d_compute(
-	poly2d * p,
-	OMPInt x,
-	OMPInt y) {
-	DFloat z = 0.00;
-	for (DLong i = 0; i < p->nc; i++) 		z+=p->c[i] * powerOptimised(x, p->px[i]) * powerOptimised(y, p->py[i]);
-	return z;
-  }
-  DFloat * poly2d_compute_init_x(poly2d * p,	SizeT n) {
-	DFloat * res= (DFloat*) malloc(p->nc*n*sizeof(DFloat));
-	for (auto i=0, s=0; i< n; ++i) for (DLong k = 0; k < p->nc; k++) res[s++]=powerOptimised(i, p->px[k]);
-	return res;
-  }
-  DFloat* poly2d_compute_init_y(poly2d * p,	SizeT n) {
-	DFloat * res= (DFloat*) malloc(p->nc*n*sizeof(DFloat));
-	for (auto i=0,s=0; i< n; ++i) for (DLong k = 0; k < p->nc; k++) res[s++]=powerOptimised(i, p->py[k]);
-	return res;
-  }
-  /*-------------------------------------------------------------------------*/
-  /**
-	@brief	Generate an interpolation kernel to use in this module.
-	@param	kernel_type (integer) 1:linear 2:cubic 3 quintic
-	@param	cubic (DFloat) cubic parameter [0..1[ for cubic kernel.
-	@return	1 newly allocated array of floats.
-
-	The returned array of floats must be deallocated using free().
-   */
-
-  /*--------------------------------------------------------------------------*/
-  DFloat * generate_interpolation_kernel(int kernel_type, DFloat cubicParameter) {
-	float * tab;
-	int i;
-	float x;
-	int samples = KERNEL_SAMPLES;
-	/*
-	  Taken from "Image Reconstruction By Piecewise Polynomial Kernels", 
-	  Meijering et al
-	 */
-
-	if (kernel_type == 1) {
-	  tab = (float *) calloc(samples, sizeof (float));
-	  tab[0] = 1.0;
-	  for (i = 1; i < samples; ++i) {
-		x = (float) KERNEL_WIDTH * (float) i / (float) (samples - 1);
-		if (x < 1)
-		  tab[i] = -x + 1;
-		else if (x >= 1)
-		  tab[i] = 0;
-	  }
-	} else if (kernel_type == 2) { //uses cubic
-	  tab = (float *) calloc(samples, sizeof (float));
-	  tab[0] = 1.0;
-	  for (i = 1; i < samples; ++i) {
-		x = (float) KERNEL_WIDTH * (float) i / (float) (samples - 1);
-		if (x < 1)
-		  tab[i] = (cubicParameter + 2) * gdl::powI(x, 3) - (cubicParameter + 3) * gdl::powI(x, 2) + 1;
-		else if (x < 2)
-		  tab[i] = cubicParameter * gdl::powI(x, 3) -
-		  (5 * cubicParameter) * gdl::powI(x, 2) + (8 * cubicParameter) * x - (4 * cubicParameter);
-	  }
-	} else if (kernel_type == 3) { //quintic
-	  tab = (float *) calloc(samples, sizeof (float));
-	  tab[0] = 1.0;
-	  for (i = 1; i < samples; ++i) {
-		x = (float) KERNEL_WIDTH * (float) i / (float) (samples - 1);
-		if (x < 1)
-		  tab[i] = (10. * cubicParameter - (21. / 16.)) * gdl::powI(x, 5) +
-		  (-18. * cubicParameter + (45. / 16)) * gdl::powI(x, 4)+
-		  (8. * cubicParameter - (5. / 2.)) * gdl::powI(x, 2) +
-		  1.0;
-		else if (x < 2)
-		  tab[i] = (11. * cubicParameter - (5. / 16.)) * gdl::powI(x, 5)+
-		  (-88. * cubicParameter + (45. / 16.)) * gdl::powI(x, 4)+
-		  (270. * cubicParameter - 10) * gdl::powI(x, 3)+
-		  (-392. * cubicParameter + (35. / 2.)) * gdl::powI(x, 2)+
-		  (265. * cubicParameter - 15.) * x +
-		  (-66. * cubicParameter + 5);
-		else if (x < 3)
-		  tab[i] = cubicParameter * gdl::powI(x, 5) +
-		  (-14. * cubicParameter) * gdl::powI(x, 4) +
-		  (78. * cubicParameter) * gdl::powI(x, 3) +
-		  (-216. * cubicParameter) * gdl::powI(x, 2) +
-		  297. * cubicParameter * x +
-		  (-162. * cubicParameter);
-	  }
-	} else {
-	  throw GDLException("Internal GDL error in generate_interpolation_kernel(), please report.");
-	  return NULL;
-	}
-	return tab;
-  }
-
-  template< typename T1, typename T2>
-  BaseGDL* warp_linear0(
-    const SizeT nCols,
-    const SizeT nRows,
-    BaseGDL* data_,
-    DFloat * const P,
-    DFloat * const Q,
-    DFloat const initvalue_,
-	const bool doMissing) {
-	std::cerr<<"warp_linear0\n";
-	SizeT lx = data_->Dim(0);
-	SizeT ly = data_->Dim(1);
-
-	dimension dim(nCols, nRows);
-	T1* res_ = new T1(dim, BaseGDL::NOZERO);
-	T2 initvalue = initvalue_;
-	SizeT nEl = nCols*nRows;
-
-	T2* res = (T2*) res_->DataAddr();
-	T2* const data = (T2* const) data_->DataAddr();
-	if (doMissing) {
-	  if ((GDL_NTHREADS = parallelize(nEl)) == 1) {
-		for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-	  } else {
-		TRACEOMP(__FILE__, __LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS)
-		  for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-	  }
-	}
-	//these are accelerators - not using them increase exec time by 2 or 3.
-	DFloat fllx=lx-1;
-	DFloat flly=ly-1;
-	SizeT llx = lx - 1;
-	SizeT lly = ly - 1;
-	DFloat p0 = P[0];
-	DFloat q0 = Q[0];
-	DFloat p1 = P[1];
-	DFloat q1 = Q[1];
-	DFloat p2 = P[2];
-	DFloat q2 = Q[2];
-	DFloat p3 = P[3];
-	DFloat q3 = Q[3];
-	DFloat p1j, p3j, q1j, q3j;
-	if (doMissing) {
-	  /* Double loop on the output image  */
-	  if ((GDL_NTHREADS = parallelize(nEl)) == 1) {
-		for (OMPInt j = 0; j < nRows; ++j) {
-		  p1j = p0 + p1 * j;
-		  p3j = p2 + p3 * j;
-		  q1j = q0 + q1 * j;
-		  q3j = q2 + q3 * j;
-		  for (OMPInt i = 0; i < nCols; ++i) {
-			// Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-			DFloat x = p1j + p3j * i;
-			if (x < 0) continue;
-			if (x >= lx) continue;
-			DFloat y = q1j + q3j * i;
-			if (y < 0) continue;
-			if (y >= ly) continue;
-			SizeT px=x;
-			SizeT py=y;
-			res[i + j * nCols] = data[px + py * lx];
-		  }
-		}
-	  } else {
-		TRACEOMP(__FILE__, __LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS) private(p1j,p3j,q1j,q3j)
-		  for (OMPInt j = 0; j < nRows; ++j) {
-		  p1j = p0 + p1 * j;
-		  p3j = p2 + p3 * j;
-		  q1j = q0 + q1 * j;
-		  q3j = q2 + q3 * j;
-		  for (OMPInt i = 0; i < nCols; ++i) {
-			// Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-		    DFloat x = p1j + p3j * i;
-			if (x < 0) continue;
-			if (x >= lx) continue;
-			DFloat y = q1j + q3j * i;
-			if (y < 0) continue;
-			if (y >= ly) continue;
-			SizeT px=x;
-			SizeT py=y;
-			res[i + j * nCols] = data[px + py * lx];
-		  }
-		}
-	  }
-	  
-	} else {
-	/* Double loop on the output image  */
-	if ((GDL_NTHREADS = parallelize(nEl)) == 1) {
-	  for (OMPInt j = 0; j < nRows; ++j) {
-		  p1j = p0 + p1 * j;
-		  p3j = p2 + p3 * j;
-		  q1j = q0 + q1 * j;
-		  q3j = q2 + q3 * j;
-		for (OMPInt i = 0; i < nCols; ++i) {
-		  // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-		  DFloat x = p1j + p3j * i;
-		  SizeT px=x;
-		  if (x < 0) px = 0; else if (x > fllx) px = llx;
-		  DFloat y = q1j + q3j * i;
-		  SizeT py=y;
-		  if (y < 0) py = 0; else if (y > flly) py = lly;
-		  res[i + j * nCols] = data[px + py * lx];
-		}
-	  }
-	} else {
-	  TRACEOMP(__FILE__, __LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS) private(p1j,p3j,q1j,q3j)
-	  for (OMPInt j = 0; j < nRows; ++j) {
-		  p1j = p0 + p1 * j;
-		  p3j = p2 + p3 * j;
-		  q1j = q0 + q1 * j;
-		  q3j = q2 + q3 * j;
-		for (OMPInt i = 0; i < nCols; ++i) {
-		  // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-		  DFloat x = p1j + p3j * i;
-		  SizeT px=x;
-		  if (x < 0) px = 0; else if (x > fllx) px = fllx;
-	      DFloat y = q1j + q3j * i;
-		  SizeT py=y;
-		  if (y < 0) py = 0; else if (y > flly) py = flly;
-		  res[i + j * nCols] = data[px + py * lx];
-		}
-	  }
-	}
-  }
-	return res_;
-  }
-
-  template< typename T1, typename T2>
-  BaseGDL* warp0(
-    const SizeT nCols,
-    const SizeT nRows,
-    BaseGDL* const data_,
-    poly2d* const poly_u,
-    poly2d* const poly_v,
-    const DFloat initvalue_,
-    bool doMissing) {
-	std::cerr<<"warp0\n";
-	const SizeT lx = data_->Dim(0);
-	const SizeT ly = data_->Dim(1);
-
-	dimension dim(nCols, nRows);
-	T1* res_ = new T1(dim, BaseGDL::NOZERO);
-	T2 initvalue = initvalue_;
-	const SizeT nEl = nCols*nRows;
-
-	T2* res = (T2*) res_->DataAddr();
-	T2* const data = (T2* const) data_->DataAddr();
-	if (doMissing) {
-	if ((GDL_NTHREADS = parallelize(nEl)) == 1) {
-		for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-	  } else {
-		TRACEOMP(__FILE__, __LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS)
-		  for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-	  }
-	}
-	//these are accelerators - not using them increase exec time by 2 or 3.
-	DFloat fllx = lx - 1;
-	DFloat flly = ly - 1;
-	SizeT llx = lx - 1;
-	SizeT lly = ly - 1;
-	DLong nc=poly_u->nc;
-	DFloat * const xcoefu=poly2d_compute_init_x(poly_u,lx);
-	DFloat * const ycoefu=poly2d_compute_init_y(poly_u,lx);
-	DFloat * const xcoefv=poly2d_compute_init_x(poly_v,ly);
-	DFloat * const ycoefv=poly2d_compute_init_y(poly_v,ly);
-	/* Double loop on the output image  */
-	if (doMissing) {
-	if ((GDL_NTHREADS = parallelize(nEl)) == 1) {
-	  for (OMPInt j = 0; j < nRows; ++j) {
-		for (OMPInt i = 0; i < nCols; ++i) {
-          DFloat x = poly_u->c[0]; for (auto k=1; k< nc; ++k) x+=poly_u->c[k]*xcoefu[j*nc+k]*ycoefu[i*nc+k];
-		  if ((x < 0) || (x > lx -1) ) continue; // already initialised to 'missing' value.
-          DFloat y = poly_v->c[0]; for (auto k=1; k< nc; ++k) y+=poly_v->c[k]*xcoefv[j*nc+k]*ycoefv[i*nc+k];
-		  if ((y < 0) || (y > ly -1)) 	continue;
-		  SizeT px = x;
-		  SizeT py = y;
-		  res[i + j * nCols] = data[px + py * lx];
-		}
-	  }
-	} else {
-	  TRACEOMP(__FILE__, __LINE__)
-#pragma omp parallel for collapse(2) num_threads(GDL_NTHREADS)
-	  for (OMPInt j = 0; j < nRows; ++j) {
-		for (OMPInt i = 0; i < nCols; ++i) {
-		  // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-          DFloat x = poly_u->c[0]; for (auto k=1; k< nc; ++k) x+=poly_u->c[k]*xcoefu[j*nc+k]*ycoefu[i*nc+k];
-		  if ((x < 0) || (x > lx -1)) continue;
-          DFloat y = poly_v->c[0]; for (auto k=1; k< nc; ++k) y+=poly_v->c[k]*xcoefv[j*nc+k]*ycoefv[i*nc+k];
-		  if ((y < 0) || (y > ly -1)) 	continue;
-		  SizeT px = x;
-		  SizeT py = y;
-		  res[i + j * nCols] = data[px + py * lx];
-		}
-	  }
-	}
-	} else {
-	/* Double loop on the output image  */
-	if ((GDL_NTHREADS = parallelize(nEl)) == 1) {
-	  for (OMPInt j = 0; j < nRows; ++j) {
-		for (OMPInt i = 0; i < nCols; ++i) {
-          DFloat x = poly_u->c[0];
-		  DFloat y = poly_v->c[0];
-		  for (auto k=1; k< nc; ++k) {
-			x+=poly_u->c[k]*xcoefu[j*nc+k]*ycoefu[i*nc+k];
-            y+=poly_v->c[k]*xcoefv[j*nc+k]*ycoefv[i*nc+k];
-		  }
-		  SizeT px=x;
-		  SizeT py=y;
-		  if (x < 0) px = 0; else if (x > fllx) px = llx;
-		  if (y < 0) py = 0; else if (y > flly) py = lly;
-		  res[i + j * nCols] = data[px + py * lx];
-		}
-	  }
-	} else {
-	  TRACEOMP(__FILE__, __LINE__)
-#pragma omp parallel for collapse(2) num_threads(GDL_NTHREADS)
-	  for (OMPInt j = 0; j < nRows; ++j) {
-		for (OMPInt i = 0; i < nCols; ++i) {
-		  // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-          DFloat x = poly_u->c[0];
-		  DFloat y = poly_v->c[0];
-		  for (auto k=1; k< nc; ++k) {
-			x+=poly_u->c[k]*xcoefu[j*nc+k]*ycoefu[i*nc+k];
-            y+=poly_v->c[k]*xcoefv[j*nc+k]*ycoefv[i*nc+k];
-		  }
-		  SizeT px=x;
-		  SizeT py=y;
-		  if (x < 0) px = 0; else if (x > fllx) px = llx;
-		  if (y < 0) py = 0; else if (y > flly) py = lly;
-		  res[i + j * nCols] = data[px + py * lx];
-		}
-	  }
-	}
-	}
-	free(xcoefu);
-	free(ycoefu);
-	free(xcoefv);
-	free(ycoefv);
-	free(poly_u->px);
-	free(poly_u->py);
-	free(poly_u->c);
-	free(poly_u);
-	free(poly_v->px);
-	free(poly_v->py);
-	free(poly_v->c);
-	free(poly_v);
-	return res_;
-  }
-
-  template< typename T1, typename T2>
-  BaseGDL* warp_linear1(
-    SizeT nCols,
-    SizeT nRows,
-    BaseGDL* data_,
-    DFloat *P,
-    DFloat *Q,
-    DFloat missingValue,
-    bool doMissing) {
-	std::cerr<<"warp_linear1\n";
-
-    SizeT lx = data_->Dim(0);
-    SizeT ly = data_->Dim(1);
-
-    dimension dim(nCols, nRows);
-    T1* res_ = new T1(dim, BaseGDL::NOZERO);
-    T2 initvalue = missingValue;
-    SizeT nEl = nCols*nRows;
-
-    T2* res = (T2*) res_->DataAddr();
-    T2* data = (T2*) data_->DataAddr();
-
-    if (doMissing) {
-      if ((GDL_NTHREADS=parallelize( nEl))==1) {
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      } else {
-      TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS)
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      }
-    }
-
-    /* Double loop on the output internal image  */
-    if ((GDL_NTHREADS=parallelize( nEl ))==1) {
-      for (OMPInt j = 0; j < nRows; ++j) {
-        for (OMPInt i = 0; i < nCols; ++i) {
-          // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-			DFloat x = P[0] + P[1] * j + P[2] * i;
-			DFloat y = Q[0] + Q[1] * j + Q[2] * i;
-          if (doMissing && ((x < 0) || (x >= lx ) || (y < 0) || (y >= ly ))) {
-            continue; // already initialised to 'missing' value.
-          } else {
-			if (x < 1) x = 1; else if (x > lx - 2) x = lx - 2;
-			if (y < 1) y = 1; else if (y > ly - 2) y = ly - 2;
-			SizeT px = x;
-			SizeT py = y;
-			SizeT pixnum[4];
-			DFloat dx = x - px;
-			DFloat dy = y - py;
-			pixnum[0]=px + py * lx;
-			pixnum[1]=pixnum[0]+1;
-			pixnum[2]=pixnum[0]+lx;
-			pixnum[3]=pixnum[2]+1;
-			DFloat a = data[pixnum[0]] + dx * (data[pixnum[1]]-data[pixnum[0]]);
-			DFloat b = data[pixnum[2]] + dx * (data[pixnum[3]]-data[pixnum[2]]);
-			res[i + j * nCols] = a + dy * (b-a);
-			//			DFloat umdx=1 - dx;
-			//			DFloat umdy=1 - dy;
-            // same time:			res[i + j * nCols] = data[pixnum[0]]*umdx*umdy+data[pixnum[1]]*dx*umdy+data[pixnum[2]]*dy*umdx+data[pixnum[3]]*dx*dy;
-            }
-        }
-	  }
-    } else {
-    TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for collapse(2) num_threads(GDL_NTHREADS)
-		for (OMPInt j = 0; j < nRows; ++j) {
-		for (OMPInt i = 0; i < nCols; ++i) {
-		  // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-		  DFloat x = P[0] + P[1] * j + P[2] * i;
-		  DFloat y = Q[0] + Q[1] * j + Q[2] * i;
-		  if (doMissing && ((x < 0) || (x >= lx) || (y < 0) || (y >= ly))) {
-			continue; // already initialised to 'missing' value.
-		  } else {
-			if (x < 1) x = 1; else if (x > lx - 2) x = lx - 2;
-			if (y < 1) y = 1; else if (y > ly - 2) y = ly - 2;
-			SizeT px = x;
-			SizeT py = y;
-			SizeT pixnum[4];
-			DFloat dx = x - px;
-			DFloat dy = y - py;
-			pixnum[0]=px + py * lx;
-			pixnum[1]=pixnum[0]+1;
-			pixnum[2]=pixnum[0]+lx;
-			pixnum[3]=pixnum[2]+1;
-			DFloat a = data[pixnum[0]] + dx * (data[pixnum[1]]-data[pixnum[0]]);
-			DFloat b = data[pixnum[2]] + dx * (data[pixnum[3]]-data[pixnum[2]]);
-			res[i + j * nCols] = a + dy * (b-a);
-			//			DFloat umdx=1 - dx;
-			//			DFloat umdy=1 - dy;
-            // same time:			res[i + j * nCols] = data[pixnum[0]]*umdx*umdy+data[pixnum[1]]*dx*umdy+data[pixnum[2]]*dy*umdx+data[pixnum[3]]*dx*dy;
-		  }
-		}
-	  }
-    }
-    return res_;
-  }
-
-  template< typename T1, typename T2>
-  BaseGDL* warp1(
-    SizeT nCols,
-    SizeT nRows,
-    BaseGDL* data_,
-    poly2d* poly_u,
-    poly2d* poly_v,
-    DFloat missingValue,
-    bool doMissing) {
-
-	std::cerr<<"warp1\n";
-
-    SizeT lx = data_->Dim(0);
-    SizeT ly = data_->Dim(1);
-
-    dimension dim(nCols, nRows);
-    T1* res_ = new T1(dim, BaseGDL::NOZERO);
-    T2 initvalue = missingValue;
-    SizeT nEl = nCols*nRows;
-
-    T2* res = (T2*) res_->DataAddr();
-    T2* data = (T2*) data_->DataAddr();
-
-    if (doMissing) {
-      if ((GDL_NTHREADS=parallelize( nEl))==1) {
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      } else {
-      TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS)
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      }
-    }
-
-    /* Double loop on the output internal image  */
-    if ((GDL_NTHREADS=parallelize( nEl ))==1) {
-      for (OMPInt j = 0; j < nRows; ++j) {
-        for (OMPInt i = 0; i < nCols; ++i) {
-          // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-          DFloat x = poly2d_compute(poly_u, j, i);
-          DFloat y = poly2d_compute(poly_v, j, i);
-          if (doMissing && ((x < 0) || (x >= lx ) || (y < 0) || (y >= ly ))) {
-            continue; // already initialised to 'missing' value.
-          } else {
-			if (x < 1) x = 1; else if (x > lx - 2) x = lx - 2;
-			if (y < 1) y = 1; else if (y > ly - 2) y = ly - 2;
-			SizeT px = x;
-			SizeT py = y;
-			SizeT pixnum[4];
-			DFloat dx = x - px;
-			DFloat dy = y - py;
-			pixnum[0]=px + py * lx;
-			pixnum[1]=pixnum[0]+1;
-			pixnum[2]=pixnum[0]+lx;
-			pixnum[3]=pixnum[2]+1;
-			DFloat a = data[pixnum[0]] + dx * (data[pixnum[1]]-data[pixnum[0]]);
-			DFloat b = data[pixnum[2]] + dx * (data[pixnum[3]]-data[pixnum[2]]);
-			res[i + j * nCols] = a + dy * (b-a);
-			//			DFloat umdx=1 - dx;
-			//			DFloat umdy=1 - dy;
-            // same time:			res[i + j * nCols] = data[pixnum[0]]*umdx*umdy+data[pixnum[1]]*dx*umdy+data[pixnum[2]]*dy*umdx+data[pixnum[3]]*dx*dy;
-            }
-        }
-	  }
-    } else {
-    TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for collapse(2) num_threads(GDL_NTHREADS)
-		for (OMPInt j = 0; j < nRows; ++j) {
-		for (OMPInt i = 0; i < nCols; ++i) {
-		  // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-          DFloat x = poly2d_compute(poly_u, j, i);
-          DFloat y = poly2d_compute(poly_v, j, i);
-		  if (doMissing && ((x < 0) || (x >= lx) || (y < 0) || (y >= ly))) {
-			continue; // already initialised to 'missing' value.
-		  } else {
-			if (x < 1) x = 1; else if (x > lx - 2) x = lx - 2;
-			if (y < 1) y = 1; else if (y > ly - 2) y = ly - 2;
-			SizeT px = x;
-			SizeT py = y;
-			SizeT pixnum[4];
-			DFloat dx = x - px;
-			DFloat dy = y - py;
-			pixnum[0]=px + py * lx;
-			pixnum[1]=pixnum[0]+1;
-			pixnum[2]=pixnum[0]+lx;
-			pixnum[3]=pixnum[2]+1;
-			DFloat a = data[pixnum[0]] + dx * (data[pixnum[1]]-data[pixnum[0]]);
-			DFloat b = data[pixnum[2]] + dx * (data[pixnum[3]]-data[pixnum[2]]);
-			res[i + j * nCols] = a + dy * (b-a);
-			//			DFloat umdx=1 - dx;
-			//			DFloat umdy=1 - dy;
-            // same time:			res[i + j * nCols] = data[pixnum[0]]*umdx*umdy+data[pixnum[1]]*dx*umdy+data[pixnum[2]]*dy*umdx+data[pixnum[3]]*dx*dy;
-		  }
-		}
-	  }
-	}
-	free(poly_u->px);
-	free(poly_u->py);
-	free(poly_u->c);
-	free(poly_u);
-	free(poly_v->px);
-	free(poly_v->py);
-	free(poly_v->c);
-	free(poly_v);
-    return res_;
-  }
-
-  template< typename T1, typename T2>
-  BaseGDL* warp_linear2(
-    SizeT nCols,
-    SizeT nRows,
-    BaseGDL* data_,
-    DFloat *P,
-    DFloat *Q,
-    DFloat cubicParameter,
-    DFloat missingValue,
-    bool doMissing) {
-	std::cerr<<"warp_linear2\n";
-
-    DLong lx = data_->Dim(0);
-    DLong ly = data_->Dim(1);
-
-    dimension dim(nCols, nRows);
-    T1* res_ = new T1(dim, BaseGDL::NOZERO);
-    T2 initvalue = missingValue;
-    DLong nEl = nCols*nRows;
-
-    T2* res = (T2*) res_->DataAddr();
-    T2* data = (T2*) data_->DataAddr();
-
-    int leaps[16];
-
-    DFloat * kernel = generate_interpolation_kernel(2, cubicParameter);
-
-    /* Pre compute leaps for 16 closest neighbors positions */
-
-    leaps[0] = -1 - lx;
-    leaps[1] = -lx;
-    leaps[2] = 1 - lx;
-    leaps[3] = 2 - lx;
-
-    leaps[4] = -1;
-    leaps[5] = 0;
-    leaps[6] = 1;
-    leaps[7] = 2;
-
-    leaps[8] = -1 + lx;
-    leaps[9] = lx;
-    leaps[10] = 1 + lx;
-    leaps[11] = 2 + lx;
-
-    leaps[12] = -1 + 2 * lx;
-    leaps[13] = 2 * lx;
-    leaps[14] = 1 + 2 * lx;
-    leaps[15] = 2 + 2 * lx;
-
-
-    if (doMissing) {
-      if ((GDL_NTHREADS=parallelize( nEl))==1) {
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      } else {
-      TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS)
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      }
-    }
-
-    /* Double loop on the output image  */
-    if ((GDL_NTHREADS=parallelize( nEl, TP_CPU_INTENSIVE))==1) {
-      for (OMPInt j = 0; j < nRows; ++j) {
-        for (OMPInt i = 0; i < nCols; ++i) {
-          // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-			DFloat x = P[0] + P[1] * j + P[2] * i;
-			DFloat y = Q[0] + Q[1] * j + Q[2] * i;
-			OMPInt px = static_cast<OMPInt> (floor(x));
-			OMPInt py = static_cast<OMPInt> (floor(y));
-          if (doMissing && ((px < 0) || (px > (lx - 1)) || (py < 0) || (py > (ly - 1)))) {
-            continue; // already initialised to 'missing' value.
-          } else {
-            if (px < 1) px = 0;
-            if (px > (lx - 1)) px = (lx - 1);
-            if (py < 1) py = 0;
-            if (py > (ly - 1)) py = (ly - 1);
-            if ((px < 1) || (px > (lx - 3)) || (py < 1) || (py > (ly - 3))) res[i + j * nCols] = data[px + py * lx];
-            else {
-              DFloat cur;
-              DFloat neighbors[16];
-              DFloat rsc[8], sumrs;
-
-              // Feed the positions for the closest 16 neighbors
-              for (SizeT k = 0; k < 16; ++k) neighbors[k] = data[px + py * lx + leaps[k]];
-
-              DFloat dpx = (DFloat) px;
-              DFloat dpy = (DFloat) py;
-              // Which tabulated value index shall we use?
-              DLong tabx = (DLong) ((x - dpx) * (DFloat) (TABSPERPIX));
-              DLong taby = (DLong) ((y - dpy) * (DFloat) (TABSPERPIX));
-
-              /* Compute resampling coefficients  */
-              /* rsc[0..3] in x, rsc[4..7] in y   */
-
-              rsc[0] = kernel[TABSPERPIX + tabx];
-              rsc[1] = kernel[tabx];
-              rsc[2] = kernel[TABSPERPIX - tabx];
-              rsc[3] = kernel[2 * TABSPERPIX - tabx];
-              rsc[4] = kernel[TABSPERPIX + taby];
-              rsc[5] = kernel[taby];
-              rsc[6] = kernel[TABSPERPIX - taby];
-              rsc[7] = kernel[2 * TABSPERPIX - taby];
-
-              sumrs = (rsc[0] + rsc[1] + rsc[2] + rsc[3]) *
-                (rsc[4] + rsc[5] + rsc[6] + rsc[7]);
-
-              /* Compute interpolated pixel now   */
-              cur = rsc[4] * (rsc[0] * neighbors[0] +
-                rsc[1] * neighbors[1] +
-                rsc[2] * neighbors[2] +
-                rsc[3] * neighbors[3]) +
-                rsc[5] * (rsc[0] * neighbors[4] +
-                rsc[1] * neighbors[5] +
-                rsc[2] * neighbors[6] +
-                rsc[3] * neighbors[7]) +
-                rsc[6] * (rsc[0] * neighbors[8] +
-                rsc[1] * neighbors[9] +
-                rsc[2] * neighbors[10] +
-                rsc[3] * neighbors[11]) +
-                rsc[7] * (rsc[0] * neighbors[12] +
-                rsc[1] * neighbors[13] +
-                rsc[2] * neighbors[14] +
-                rsc[3] * neighbors[15]);
-
-              /* Affect the value to the output image */
-              res[i + j * nCols] = (cur / sumrs);
-              /* done ! */
-            }
-          }
-        }
-      }
-    } else {
-    TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for collapse(2) num_threads(GDL_NTHREADS)
-      for (OMPInt j = 0; j < nRows; ++j) {
-        for (OMPInt i = 0; i < nCols; ++i) {
-          // Compute the original source for this pixel, note order of j and i in P and Q definition of IDL doc.
-			DFloat x = P[0] + P[1] * j + P[2] * i;
-			DFloat y = Q[0] + Q[1] * j + Q[2] * i;
-			OMPInt px = static_cast<OMPInt> (floor(x));
-			OMPInt py = static_cast<OMPInt> (floor(y));
-          if (doMissing && ((px < 0) || (px > (lx - 1)) || (py < 0) || (py > (ly - 1)))) {
-            continue; // already initialised to 'missing' value.
-          } else {
-            if (px < 1) px = 0;
-            if (px > (lx - 1)) px = (lx - 1);
-            if (py < 1) py = 0;
-            if (py > (ly - 1)) py = (ly - 1);
-            if ((px < 1) || (px > (lx - 3)) || (py < 1) || (py > (ly - 3))) res[i + j * nCols] = data[px + py * lx];
-            else {
-              DFloat cur;
-              DFloat neighbors[16];
-              DFloat rsc[8], sumrs;
-
-              // Feed the positions for the closest 16 neighbors
-              for (SizeT k = 0; k < 16; ++k) neighbors[k] = data[px + py * lx + leaps[k]];
-
-              DFloat dpx = (DFloat) px;
-              DFloat dpy = (DFloat) py;
-              // Which tabulated value index shall we use?
-              DLong tabx = (DLong) ((x - dpx) * (DFloat) (TABSPERPIX));
-              DLong taby = (DLong) ((y - dpy) * (DFloat) (TABSPERPIX));
-
-              /* Compute resampling coefficients  */
-              /* rsc[0..3] in x, rsc[4..7] in y   */
-
-              rsc[0] = kernel[TABSPERPIX + tabx];
-              rsc[1] = kernel[tabx];
-              rsc[2] = kernel[TABSPERPIX - tabx];
-              rsc[3] = kernel[2 * TABSPERPIX - tabx];
-              rsc[4] = kernel[TABSPERPIX + taby];
-              rsc[5] = kernel[taby];
-              rsc[6] = kernel[TABSPERPIX - taby];
-              rsc[7] = kernel[2 * TABSPERPIX - taby];
-
-              sumrs = (rsc[0] + rsc[1] + rsc[2] + rsc[3]) *
-                (rsc[4] + rsc[5] + rsc[6] + rsc[7]);
-
-              /* Compute interpolated pixel now   */
-              cur = rsc[4] * (rsc[0] * neighbors[0] +
-                rsc[1] * neighbors[1] +
-                rsc[2] * neighbors[2] +
-                rsc[3] * neighbors[3]) +
-                rsc[5] * (rsc[0] * neighbors[4] +
-                rsc[1] * neighbors[5] +
-                rsc[2] * neighbors[6] +
-                rsc[3] * neighbors[7]) +
-                rsc[6] * (rsc[0] * neighbors[8] +
-                rsc[1] * neighbors[9] +
-                rsc[2] * neighbors[10] +
-                rsc[3] * neighbors[11]) +
-                rsc[7] * (rsc[0] * neighbors[12] +
-                rsc[1] * neighbors[13] +
-                rsc[2] * neighbors[14] +
-                rsc[3] * neighbors[15]);
-
-              /* Affect the value to the output image */
-              res[i + j * nCols] = (cur / sumrs);
-              /* done ! */
-            }
-          }
-        }
-      }
-    }
-    free(kernel);
-    return res_;
-  }
-
-  template< typename T1, typename T2>
-  BaseGDL* warp2(
-    SizeT nCols,
-    SizeT nRows,
-    BaseGDL* data_,
-    DFloat cubicParameter,
-    poly2d* poly_u,
-    poly2d* poly_v,
-    DFloat missingValue,
-    bool doMissing) {
-
-	std::cerr<<"warp2\n";
-    DLong lx = data_->Dim(0);
-    DLong ly = data_->Dim(1);
-
-    dimension dim(nCols, nRows);
-    T1* res_ = new T1(dim, BaseGDL::NOZERO);
-    T2 initvalue = missingValue;
-    DLong nEl = nCols*nRows;
-
-    T2* res = (T2*) res_->DataAddr();
-    T2* data = (T2*) data_->DataAddr();
-
-    int leaps[16];
-
-    DFloat * kernel = generate_interpolation_kernel(2, cubicParameter);
-
-    /* Pre compute leaps for 16 closest neighbors positions */
-
-    leaps[0] = -1 - lx;
-    leaps[1] = -lx;
-    leaps[2] = 1 - lx;
-    leaps[3] = 2 - lx;
-
-    leaps[4] = -1;
-    leaps[5] = 0;
-    leaps[6] = 1;
-    leaps[7] = 2;
-
-    leaps[8] = -1 + lx;
-    leaps[9] = lx;
-    leaps[10] = 1 + lx;
-    leaps[11] = 2 + lx;
-
-    leaps[12] = -1 + 2 * lx;
-    leaps[13] = 2 * lx;
-    leaps[14] = 1 + 2 * lx;
-    leaps[15] = 2 + 2 * lx;
-
-
-    if (doMissing) {
-      if ((GDL_NTHREADS=parallelize( nEl))==1) {
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      } else {
-      TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for num_threads(GDL_NTHREADS)
-        for (OMPInt i = 0; i < nCols * nRows; ++i) res[i] = initvalue;
-      }
-    }
-
-    /* Double loop on the output image  */
-    if ((GDL_NTHREADS=parallelize( nEl, TP_CPU_INTENSIVE))==1) {
-      for (OMPInt j = 0; j < nRows; ++j) {
-        for (OMPInt i = 0; i < nCols; ++i) {
-          // Compute the original source for this pixel, note order of j and i.
-          DFloat x = poly2d_compute(poly_u, j, i);
-          DFloat y = poly2d_compute(poly_v, j, i);
-          DLong px = (DLong) x;
-          DLong py = (DLong) y;
-          if (doMissing && ((px < 0) || (px > (lx - 1)) || (py < 0) || (py > (ly - 1)))) {
-            continue; // already initialised to 'missing' value.
-          } else {
-            if (px < 1) px = 0;
-            if (px > (lx - 1)) px = (lx - 1);
-            if (py < 1) py = 0;
-            if (py > (ly - 1)) py = (ly - 1);
-            if ((px < 1) || (px > (lx - 3)) || (py < 1) || (py > (ly - 3))) res[i + j * nCols] = data[px + py * lx];
-            else {
-              DFloat cur;
-              DFloat neighbors[16];
-              DFloat rsc[8], sumrs;
-
-              // Feed the positions for the closest 16 neighbors
-              for (SizeT k = 0; k < 16; ++k) neighbors[k] = data[px + py * lx + leaps[k]];
-
-              DFloat dpx = (DFloat) px;
-              DFloat dpy = (DFloat) py;
-              // Which tabulated value index shall we use?
-              DLong tabx = (DLong) ((x - dpx) * (DFloat) (TABSPERPIX));
-              DLong taby = (DLong) ((y - dpy) * (DFloat) (TABSPERPIX));
-
-              /* Compute resampling coefficients  */
-              /* rsc[0..3] in x, rsc[4..7] in y   */
-
-              rsc[0] = kernel[TABSPERPIX + tabx];
-              rsc[1] = kernel[tabx];
-              rsc[2] = kernel[TABSPERPIX - tabx];
-              rsc[3] = kernel[2 * TABSPERPIX - tabx];
-              rsc[4] = kernel[TABSPERPIX + taby];
-              rsc[5] = kernel[taby];
-              rsc[6] = kernel[TABSPERPIX - taby];
-              rsc[7] = kernel[2 * TABSPERPIX - taby];
-
-              sumrs = (rsc[0] + rsc[1] + rsc[2] + rsc[3]) *
-                (rsc[4] + rsc[5] + rsc[6] + rsc[7]);
-
-              /* Compute interpolated pixel now   */
-              cur = rsc[4] * (rsc[0] * neighbors[0] +
-                rsc[1] * neighbors[1] +
-                rsc[2] * neighbors[2] +
-                rsc[3] * neighbors[3]) +
-                rsc[5] * (rsc[0] * neighbors[4] +
-                rsc[1] * neighbors[5] +
-                rsc[2] * neighbors[6] +
-                rsc[3] * neighbors[7]) +
-                rsc[6] * (rsc[0] * neighbors[8] +
-                rsc[1] * neighbors[9] +
-                rsc[2] * neighbors[10] +
-                rsc[3] * neighbors[11]) +
-                rsc[7] * (rsc[0] * neighbors[12] +
-                rsc[1] * neighbors[13] +
-                rsc[2] * neighbors[14] +
-                rsc[3] * neighbors[15]);
-
-              /* Affect the value to the output image */
-              res[i + j * nCols] = (cur / sumrs);
-              /* done ! */
-            }
-          }
-        }
-      }
-    } else {
-    TRACEOMP(__FILE__,__LINE__)
-#pragma omp parallel for collapse(2) num_threads(GDL_NTHREADS)
-      for (OMPInt j = 0; j < nRows; ++j) {
-        for (OMPInt i = 0; i < nCols; ++i) {
-          // Compute the original source for this pixel, note order of j and i.
-          DFloat x = poly2d_compute(poly_u,  j, i);
-          DFloat y = poly2d_compute(poly_v,  j, i);
-          DLong px = (DLong) x;
-          DLong py = (DLong) y;
-          if (doMissing && ((px < 0) || (px > (lx - 1)) || (py < 0) || (py > (ly - 1)))) {
-            continue; // already initialised to 'missing' value.
-          } else {
-            if (px < 1) px = 0;
-            if (px > (lx - 1)) px = (lx - 1);
-            if (py < 1) py = 0;
-            if (py > (ly - 1)) py = (ly - 1);
-            if ((px < 1) || (px > (lx - 3)) || (py < 1) || (py > (ly - 3))) res[i + j * nCols] = data[px + py * lx];
-            else {
-              DFloat cur;
-              DFloat neighbors[16];
-              DFloat rsc[8], sumrs;
-
-              // Feed the positions for the closest 16 neighbors
-              for (SizeT k = 0; k < 16; ++k) neighbors[k] = data[px + py * lx + leaps[k]];
-
-              DFloat dpx = (DFloat) px;
-              DFloat dpy = (DFloat) py;
-              // Which tabulated value index shall we use?
-              DLong tabx = (DLong) ((x - dpx) * (DFloat) (TABSPERPIX));
-              DLong taby = (DLong) ((y - dpy) * (DFloat) (TABSPERPIX));
-
-              /* Compute resampling coefficients  */
-              /* rsc[0..3] in x, rsc[4..7] in y   */
-
-              rsc[0] = kernel[TABSPERPIX + tabx];
-              rsc[1] = kernel[tabx];
-              rsc[2] = kernel[TABSPERPIX - tabx];
-              rsc[3] = kernel[2 * TABSPERPIX - tabx];
-              rsc[4] = kernel[TABSPERPIX + taby];
-              rsc[5] = kernel[taby];
-              rsc[6] = kernel[TABSPERPIX - taby];
-              rsc[7] = kernel[2 * TABSPERPIX - taby];
-
-              sumrs = (rsc[0] + rsc[1] + rsc[2] + rsc[3]) *
-                (rsc[4] + rsc[5] + rsc[6] + rsc[7]);
-
-              /* Compute interpolated pixel now   */
-              cur = rsc[4] * (rsc[0] * neighbors[0] +
-                rsc[1] * neighbors[1] +
-                rsc[2] * neighbors[2] +
-                rsc[3] * neighbors[3]) +
-                rsc[5] * (rsc[0] * neighbors[4] +
-                rsc[1] * neighbors[5] +
-                rsc[2] * neighbors[6] +
-                rsc[3] * neighbors[7]) +
-                rsc[6] * (rsc[0] * neighbors[8] +
-                rsc[1] * neighbors[9] +
-                rsc[2] * neighbors[10] +
-                rsc[3] * neighbors[11]) +
-                rsc[7] * (rsc[0] * neighbors[12] +
-                rsc[1] * neighbors[13] +
-                rsc[2] * neighbors[14] +
-                rsc[3] * neighbors[15]);
-
-              /* Affect the value to the output image */
-              res[i + j * nCols] = (cur / sumrs);
-              /* done ! */
-            }
-          }
-        }
-      }
-    }
-    free(kernel);
-    free(poly_u->px);
-    free(poly_u->py);
-    free(poly_u->c);
-    free(poly_u);
-    free(poly_v->px);
-    free(poly_v->py);
-    free(poly_v->c);
-    free(poly_v);
-    return res_;
-  }
-
-
-  BaseGDL* poly_2d_fun(EnvT* e) {
-    /*
-    The poly_2d code is inspired by "ECLIPSE", the  ESO C Library for an Image Processing Software Environment
-    N. Devillard, "The eclipse software", The messenger No 87 - March 1997 http://www.eso.org/projects/aot/eclipse/
-     */
-
-    // GD: POLY_2D IS 5 times slower that IDL's. TBC . I note that the loops contains a lot of 'ifs', so algorithm is not efficient.
-    
-    SizeT nParam = e->NParam(3);
-
-    BaseGDL* p0 = e->GetParDefined(0);
-    
-    if (p0->Type() == GDL_COMPLEX || p0->Type() == GDL_COMPLEXDBL)  e->Throw("Complex values not supported (FIXME)");
-    
-    if (p0->Rank() != 2)
-      e->Throw("Array must have 2 dimensions: " + e->GetParString(0));
-
-    BaseGDL* p1 = e->GetParDefined(1);
-    BaseGDL* p2 = e->GetParDefined(2);
-
-    DLong interp = 0;
-    if (nParam >= 4) e->AssureLongScalarPar(3, interp);
-    if (interp < 0 || interp > 2) e->Throw("Value of Interpolation type is out of allowed range.");
-
-    if (nParam == 5)
-      e->Throw("Incorrect number of arguments.");
-
-    DLong nCol = p0->Dim(0);
-    DLong nRow = p0->Dim(1);
-    if (nParam >= 6) {
-      e->AssureLongScalarPar(4, nCol);
-      e->AssureLongScalarPar(5, nRow);
-      if (nCol <1 || nRow <1) e->Throw("Array dimensions must be greater than 0.");
-    }
-    static int CUBICIx = e->KeywordIx("CUBIC");
-    DFloat cubicParameter = -0.5;
-    if (e->KeywordSet(CUBICIx)) {
-      e->AssureFloatScalarKWIfPresent(CUBICIx, cubicParameter);
-      interp = 2;
-    }
-
-    static int MISSINGIx = e->KeywordIx("MISSING");
-    DFloat missing = 0.0;
-    bool doMissing = (e->KeywordPresent(MISSINGIx));
-    if (doMissing) {
-      e->AssureFloatScalarKWIfPresent(MISSINGIx, missing);
-    }
-	
-    static int PIXELIx = e->KeywordIx("PIXEL_CENTER");
-    bool doPix = (e->KeywordPresent(PIXELIx));
-	DFloat pixcenter = 0.0;
-	if (doPix) {
-	  e->AssureFloatScalarKWIfPresent(PIXELIx, pixcenter);
-	}
-	
-	// check P dimension first
-    DLong Psize=p1->N_Elements();
-    if (Psize < 4) e->Throw("Value of Polynomial degree is out of allowed range.");
-    
-    DFloat dPdim = sqrt((DFloat) Psize);
-    DLong nDegree = (DLong) dPdim -1;
-    DLong nc=(nDegree + 1) * (nDegree + 1);
-    if ( p2->N_Elements() < nc ) e->Throw("Coefficient arrays must have (degree+1)^2 elements");
-
-    DFloatGDL* P = static_cast<DFloatGDL*>
-      (p1->Convert2(GDL_FLOAT, BaseGDL::COPY));
-
-    DFloatGDL* Q = static_cast<DFloatGDL*>
-      (p2->Convert2(GDL_FLOAT, BaseGDL::COPY));
-	(*P)[0]+=pixcenter;
-	(*Q)[0]+=pixcenter;
-
-    if (nDegree == 1 ) { //&& (*P)[3] == 0 && (*Q)[3] == 0 ) { //LINEAR CASE
-      //return p0 if identity.
-      if ((*P)[0] == 0 && (*P)[1] == 0 && (*P)[2] == 1 && (*P)[3] == 0 && (*Q)[0] == 0 && (*Q)[1] == 1 && (*Q)[2] == 0 && (*Q)[3] == 0) {
-        return p0->Dup();
-      }
-      if (interp==0) {
-	if (p0->Type() == GDL_BYTE) {
-          return warp_linear0< DByteGDL, DByte>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_INT) {
-          return warp_linear0< DIntGDL, DInt>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_UINT) {
-          return warp_linear0< DUIntGDL, DUInt>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_LONG) {
-          return warp_linear0< DLongGDL, DLong>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_ULONG) {
-          return warp_linear0< DULongGDL, DULong>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_LONG64) {
-          return warp_linear0< DLong64GDL, DLong64>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_ULONG64) {
-          return warp_linear0< DULong64GDL, DULong64>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_FLOAT) {
-          return warp_linear0< DFloatGDL, DFloat>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_DOUBLE) {
-          return warp_linear0< DDoubleGDL, DDouble>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        }
-      } else if (interp==1) {
-         if (p0->Type() == GDL_BYTE) {
-          return warp_linear1< DByteGDL, DByte>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_INT) {
-          return warp_linear1< DIntGDL, DInt>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_UINT) {
-          return warp_linear1< DUIntGDL, DUInt>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_LONG) {
-          return warp_linear1< DLongGDL, DLong>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_ULONG) {
-          return warp_linear1< DULongGDL, DULong>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_LONG64) {
-          return warp_linear1< DLong64GDL, DLong64>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_ULONG64) {
-          return warp_linear1< DULong64GDL, DULong64>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_FLOAT) {
-          return warp_linear1< DFloatGDL, DFloat>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        } else if (p0->Type() == GDL_DOUBLE) {
-          return warp_linear1< DDoubleGDL, DDouble>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), missing, doMissing);
-        }
-      } else if (interp==2) {
-         if (p0->Type() == GDL_BYTE) {
-          return warp_linear2< DByteGDL, DByte>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_INT) {
-          return warp_linear2< DIntGDL, DInt>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_UINT) {
-          return warp_linear2< DUIntGDL, DUInt>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_LONG) {
-          return warp_linear2< DLongGDL, DLong>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_ULONG) {
-          return warp_linear2< DULongGDL, DULong>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_LONG64) {
-          return warp_linear2< DLong64GDL, DLong64>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_ULONG64) {
-          return warp_linear2< DULong64GDL, DULong64>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_FLOAT) {
-          return warp_linear2< DFloatGDL, DFloat>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        } else if (p0->Type() == GDL_DOUBLE) {
-          return warp_linear2< DDoubleGDL, DDouble>(nCol, nRow, p0, (DFloat*) P->DataAddr(),(DFloat*) Q->DataAddr(), cubicParameter, missing, doMissing);
-        }
-      }
-    } 
-	
-	else {
-    
-    //NON-LINEAR Polynomial
-
-    poly2d* poly_u;
-    poly2d* poly_v;
-
-    poly_u = (poly2d *) malloc(sizeof (poly2d));
-    poly_u->nc = nc;
-    poly_u->px = (DLong *) malloc(nc * sizeof (DLong));
-    poly_u->py = (DLong *) malloc(nc * sizeof (DLong));
-    poly_u->c = (DFloat *) malloc(nc * sizeof (DFloat));
-
-    for (SizeT i = 0; i < (nDegree + 1)*(nDegree + 1); ++i) {
-      poly_u->px[i] = i / (nDegree + 1);
-      poly_u->py[i] = i - (poly_u->px[i] * (nDegree + 1));
-      poly_u->c[i] = (*P)[poly_u->px[i]+(nDegree + 1) * poly_u->py[i]];
-    }
-
-    poly_v = (poly2d *) malloc(sizeof (poly2d));
-    poly_v->nc = nc;
-    poly_v->px = (DLong *) malloc(nc * sizeof (DLong));
-    poly_v->py = (DLong *) malloc(nc * sizeof (DLong));
-    poly_v->c = (DFloat *) malloc(nc * sizeof (DFloat));
-
-    for (SizeT i = 0; i < (nDegree + 1)*(nDegree + 1); ++i) {
-      poly_v->px[i] = i / (nDegree + 1);
-      poly_v->py[i] = i - (poly_v->px[i] * (nDegree + 1));
-      poly_v->c[i] = (*Q)[poly_v->px[i]+(nDegree + 1) * poly_v->py[i]];
-    }
-    if (interp==0) {
-       if (p0->Type() == GDL_BYTE) {
-        return warp0< DByteGDL, DByte>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_INT) {
-        return warp0< DIntGDL, DInt>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_UINT) {
-        return warp0< DUIntGDL, DUInt>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_LONG) {
-        return warp0< DLongGDL, DLong>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_ULONG) {
-        return warp0< DULongGDL, DULong>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_LONG64) {
-        return warp0< DLong64GDL, DLong64>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_ULONG64) {
-        return warp0< DULong64GDL, DULong64>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_FLOAT) {
-        return warp0< DFloatGDL, DFloat>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_DOUBLE) {
-        return warp0< DDoubleGDL, DDouble>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      }
-    } else if (interp==1) {
-       if (p0->Type() == GDL_BYTE) {
-        return warp1< DByteGDL, DByte>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_INT) {
-        return warp1< DIntGDL, DInt>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_UINT) {
-        return warp1< DUIntGDL, DUInt>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_LONG) {
-        return warp1< DLongGDL, DLong>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_ULONG) {
-        return warp1< DULongGDL, DULong>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_LONG64) {
-        return warp1< DLong64GDL, DLong64>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_ULONG64) {
-        return warp1< DULong64GDL, DULong64>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_FLOAT) {
-        return warp1< DFloatGDL, DFloat>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_DOUBLE) {
-        return warp1< DDoubleGDL, DDouble>(nCol, nRow, p0, poly_u, poly_v, missing, doMissing);
-      }
-    } else if (interp==2) {
-       if (p0->Type() == GDL_BYTE) {
-        return warp2< DByteGDL, DByte>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_INT) {
-        return warp2< DIntGDL, DInt>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_UINT) {
-        return warp2< DUIntGDL, DUInt>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_LONG) {
-        return warp2< DLongGDL, DLong>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_ULONG) {
-        return warp2< DULongGDL, DULong>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_LONG64) {
-        return warp2< DLong64GDL, DLong64>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_ULONG64) {
-        return warp2< DULong64GDL, DULong64>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_FLOAT) {
-        return warp2< DFloatGDL, DFloat>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      } else if (p0->Type() == GDL_DOUBLE) {
-        return warp2< DDoubleGDL, DDouble>(nCol, nRow, p0, cubicParameter, poly_u, poly_v, missing, doMissing);
-      }
-    }
-
-    }
-    return NULL;
-  }
 
 } // namespace
